Faster code execution (#8)
* nuclei center without tile reference

* automatically determine the tile ROI

* Added reference to the WSI for nuclick

* nuclick classification returns annotation file instead of classes

* removed unwanted lines of code required for the classification part

* added label to the nuclick classification.

* added onclick functionality to SAM

* added mobile SAM model

* resolved build issues

* nuclass, nucseg, samMobile are pre-loaded

* Choose device

* moved ROI to the AI model area

* Download AI models on startup

* system used downloaded AI models

* changed file locations, updated utils and Docker

* docker issue resolved

* resolved issue with tile size calculation.

* added ai model type parameter to the UI

* prints request number for reference.

* region calculation from min nuclei area.

* corrected documentation

* moved all functionalities to utils.
