+++ .gitignore
@@ -2,3 +2,6 @@ __pycache__/
 aiInferenceModel/__pycache__
 debug
 pre-commit.sh
+aiInferenceModel/model_weights/
+.vscode
+.github
