+++ aiInferenceModel/main.py
@@ -1,10 +1,8 @@
 # Import AI model modules
-import nuclickClassification
-import nuclickSegmentation
-import samSegmentation
-import samOnclick
-import samMobile
+from ai_models import (nuclickClassification, nuclickSegmentation, samMobile,
+                       samOnclick, samSegmentation)
 from fastapi import FastAPI, HTTPException, Request
+from utils import pre_load_ai_models
 
 # You can import your custom AI models into this code
 # Example: import customSegmentationModel
@@ -12,18 +10,29 @@ from fastapi import FastAPI, HTTPException, Request
 
 app = FastAPI()
 
+# Pre-load ai models for faster execution
+mobile_sam, nuclick_class, nuclick_seg = pre_load_ai_models()
+
+# Request count
+request_count = 0
+
 
 @app.get("/")
 def read_root():
     return {"message": "DSA AI adapter for deploying AI models"}
 
 # Endpoint for Nuclick AI model classification
+
+
 @app.post("/nuclick_classification/")
 async def process_ima(request: Request):
+    global request_count
+    print(f"Request number : {request_count}")
+    request_count += 1
     try:
         json_data = await request.json()
         network_output = nuclickClassification.run_ai_model_inferencing(
-            json_data)
+            json_data, nuclick_class)
         return {"network_output": network_output}
 
     except Exception as e:
@@ -31,12 +40,17 @@ async def process_ima(request: Request):
         raise HTTPException(status_code=400, detail="Invalid data")
 
 # Endpoint for Nuclick AI model segmentation
+
+
 @app.post("/nuclick_segmentation/")
 async def process_ima(request: Request):
+    global request_count
+    print(f"Request number : {request_count}")
+    request_count += 1
     try:
         json_data = await request.json()
         network_output = nuclickSegmentation.run_ai_model_inferencing(
-            json_data)
+            json_data, nuclick_seg)
         return {"network_output": network_output}
 
     except Exception as e:
@@ -44,8 +58,13 @@ async def process_ima(request: Request):
         raise HTTPException(status_code=400, detail="Invalid data")
 
 # Endpoint for Segment anything AI model segmentation
+
+
 @app.post("/segment_anything/")
 async def process_ima(request: Request):
+    global request_count
+    print(f"Request number : {request_count}")
+    request_count += 1
     try:
         json_data = await request.json()
         network_output = samSegmentation.run_ai_model_inferencing(json_data)
@@ -54,10 +73,15 @@ async def process_ima(request: Request):
     except Exception as e:
         print(e)
         raise HTTPException(status_code=400, detail="Invalid data")
-    
+
 # Endpoint for Segment anything AI model with user input
+
+
 @app.post("/segment_anything_onclick/")
 async def process_ima(request: Request):
+    global request_count
+    print(f"Request number : {request_count}")
+    request_count += 1
     try:
         json_data = await request.json()
         network_output = samOnclick.run_ai_model_inferencing(json_data)
@@ -66,13 +90,20 @@ async def process_ima(request: Request):
     except Exception as e:
         print(e)
         raise HTTPException(status_code=400, detail="Invalid data")
-    
-# Endpoint for Mobile segment anything model    
+
+# Endpoint for Mobile segment anything model
+
+
 @app.post("/segment_anything_mobile/")
 async def process_ima(request: Request):
+    global request_count
+    print(f"Request number : {request_count}")
+    request_count += 1
     try:
         json_data = await request.json()
-        network_output = samMobile.run_ai_model_inferencing(json_data)
+        network_output = samMobile.run_ai_model_inferencing(
+            json_data, mobile_sam)
+
         return {"network_output": network_output}
 
     except Exception as e:
