+++ aiInferenceModel/requirements.txt
@@ -13,8 +13,8 @@ matplotlib
 scikit-image
 
 #segment anything
-#git+https://github.com/facebookresearch/segment-anything.git
+git+https://github.com/facebookresearch/segment-anything.git
 
 #mobile segment anything
-#git+https://github.com/ChaoningZhang/MobileSAM.git
+git+https://github.com/ChaoningZhang/MobileSAM.git
 timm
